v1.0
You see the application called entropy, where I combined the powerful features of Swift UI with the concept of cyber-security.  Today's one of the most important problems is that company employees or people do not choose a strong password.

As a result, attackers can crack weak passwords with techniques such as dictionary attacks and brute force attacks. In such a scenario, both companies and individuals may experience permanent losses.

With the swift playground called Entropy, I aimed to create cyber security awareness for users. By experiencing how strong a password can be created and how long will it take to be cracked, the user will understand the importance of the password and increase the security level.